requirements.txt
# Core dependencies
aiogram==3.13.1
aiohttp==3.10.10
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
aiomysql==0.2.0
alembic==1.13.3
redis==5.2.0

# OCR and Image Processing
Pillow==10.4.0
opencv-python==4.10.0.84
pytesseract==0.3.13
easyocr==1.7.2
google-cloud-vision==3.10.2
openai==1.90.0

# Document Processing
pypdf==5.1.0
pdf2image==1.17.0
python-docx==1.1.2
python-magic==0.4.27

# Data Processing
pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0

# Export Formats
openpyxl==3.1.5
reportlab==4.2.5
matplotlib==3.9.2
seaborn==0.13.2

# AWS Services
boto3==1.35.36
aioboto3==13.2.0

# Currency Exchange APIs
aiohttp-retry==2.8.3

# Localization
babel==2.16.0
pytz==2024.2

# Utilities
aiofiles==24.1.0
pydantic==2.9.2
pydantic-settings==2.6.0
validators==0.34.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.11.2

# Monitoring
structlog==24.4.0
prometheus-client==0.21.0

# Additional utilities
cachetools==5.5.0

# Hot reload functionality
watchdog==3.0.0